# OPENAPI/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Paperless-AI API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the Paperless-AI application",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "Clusterzx",
      "url": "https://github.com/Clusterzx"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT-based authentication for web app users. The token is obtained by authenticating via the login endpoint.\n\n### How to authenticate:\n1. Send a POST request to `/login` with your username and password\n2. The server will respond with a JWT token (also set as a cookie in browsers)\n3. Include this token in the `Authorization` header as `Bearer {token}`\n\nExample:\n```\nAuthorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\n```\n\nJWT tokens are valid for 24 hours after issuance.\n"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key authentication for programmatic access. The API key can be generated or regenerated using the /api/key-regenerate endpoint.\n\n### How to authenticate:\n1. Access the API key from your application settings\n2. Include the API key in the `x-api-key` HTTP header for all requests\n\nExample:\n```\nx-api-key: 7c1f3f5e2b0a9d8c6e4b2a1d3f5e8c9b2a1d3f5e\n```\n\nAPI keys do not expire unless regenerated.\n"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message",
            "example": "Error resetting documents"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/history": {
      "get": {
        "summary": "Document history page",
        "description": "Renders the document history page with filtering options.\nThis page displays a list of all documents that have been processed by Paperless-AI,\nshowing the changes made to the documents through AI processing.\n\nThe page includes filtering capabilities by correspondent, tag, and free text search,\nallowing users to easily find specific documents or categories of processed documents.\nEach entry includes links to the original document in Paperless-ngx.\n",
        "tags": ["History", "Navigation"],
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "History page rendered successfully",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "description": "HTML content of the history page with filtering controls and document list"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - user not authenticated",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string",
                  "example": "/login"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/history": {
      "get": {
        "summary": "Get processed document history",
        "description": "Returns a paginated list of documents that have been processed by Paperless-AI.\nSupports filtering by tag, correspondent, and search term.\nDesigned for integration with DataTables jQuery plugin.\n\nThis endpoint provides comprehensive information about each processed document,\nincluding its metadata before and after AI processing, allowing users to track\nchanges made by the system.\n",
        "tags": ["History", "API"],
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "draw",
            "schema": {
              "type": "integer"
            },
            "description": "Draw counter for DataTables (prevents XSS)",
            "example": 1
          },
          {
            "in": "query",
            "name": "start",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Starting record index for pagination",
            "example": 0
          },
          {
            "in": "query",
            "name": "length",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of records to return per page",
            "example": 10
          },
          {
            "in": "query",
            "name": "search[value]",
            "schema": {
              "type": "string"
            },
            "description": "Global search term (searches title, correspondent and tags)",
            "example": "invoice"
          },
          {
            "in": "query",
            "name": "tag",
            "schema": {
              "type": "string"
            },
            "description": "Filter by tag ID",
            "example": "5"
          },
          {
            "in": "query",
            "name": "correspondent",
            "schema": {
              "type": "string"
            },
            "description": "Filter by correspondent name",
            "example": "Acme Corp"
          },
          {
            "in": "query",
            "name": "order[0][column]",
            "schema": {
              "type": "integer"
            },
            "description": "Index of column to sort by (0=document_id, 1=title, etc.)",
            "example": 1
          },
          {
            "in": "query",
            "name": "order[0][dir]",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"]
            },
            "description": "Sort direction (ascending or descending)",
            "example": "desc"
          }
        ],
        "responses": {
          "200": {
            "description": "Document history returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "draw": {
                      "type": "integer",
                      "description": "Echo of the draw parameter",
                      "example": 1
                    },
                    "recordsTotal": {
                      "type": "integer",
                      "description": "Total number of records in the database",
                      "example": 100
                    },
                    "recordsFiltered": {
                      "type": "integer",
                      "description": "Number of records after filtering",
                      "example": 20
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "document_id": {
                            "type": "integer",
                            "description": "Document ID",
                            "example": 123
                          },
                          "title": {
                            "type": "string",
                            "description": "Document title",
                            "example": "Invoice #12345"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time",
                            "description": "Date and time when the processing occurred",
                            "example": "2023-07-15T14:30:45Z"
                          },
                          "tags": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "integer",
                                  "example": 5
                                },
                                "name": {
                                  "type": "string",
                                  "example": "Invoice"
                                },
                                "color": {
                                  "type": "string",
                                  "example": "#FF5733"
                                }
                              }
                            }
                          },
                          "correspondent": {
                            "type": "string",
                            "description": "Document correspondent name",
                            "example": "Acme Corp"
                          },
                          "link": {
                            "type": "string",
                            "description": "Link to the document in Paperless-ngx",
                            "example": "http://paperless.example.com/documents/123/"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Authentication required"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Error loading history data"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}